feat: upgrading poi libs version to 5.2.3

+++ src/main/java/org/concordion/ext/excel/conversion/cellcontent/DefaultStyleConverter.java
@@ -133,7 +133,7 @@ public class DefaultStyleConverter extends AbstractConversionStrategy<Cell> {
 	
 	protected Font getFont(Cell c) {
 		CellStyle style = c.getCellStyle();
-		short fontIndex = style.getFontIndex();
+		int fontIndex = style.getFontIndex();
 		Sheet s = c.getSheet();
 		Font f = s.getWorkbook().getFontAt(fontIndex);
 		return f;
